Source Codes/dashboardManagement/dashboardManagement/dashboardManagementRepo/dashboard_creation/requirements.txt
flask==2.3.0
mysql-connector-python-rf
requests
pandas
plotly